Add the generation of reduced compact model after CG phase
This reduced model has much fewer arcs than the one that would be
generated directly from the instance. The idea is that the RMP produces
an optimal solution for the relaxed problem, and frequently solving the
RMP as a binary program lead to feasibility issues, and long solution
times to produce a solution.

This reduced model considers only the subset of arcs present in the
optimal RMP solution, and generally is easier to solve than the binary
version of the RMP.

From my short experiments, this approach is *faster than the truncated CG
of Pepin et al. (2009), and produces better solution values.

* Still proving the optimal value for the reduced compact model is hard,
  to the user needs to set an acceptable time limit for this phase of
  the optimization.